+++ src/main/java/fi/jubic/easyvalue/processor/PropertyKind.java
@@ -1,0 +1,9 @@
+package fi.jubic.easyvalue.processor;
+
+public enum PropertyKind {
+    PLAIN,
+    OPTIONAL,
+    ARRAY,
+    OPTIONAL_ARRAY,
+    LIST;
+}