+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres poorly to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Wrap `List` instances in `Collections::unmodifiableList` to appease spotbugs.
 
 ## [1.3.3] - 2021-12-10
 ### Added